requirements.txt
# Big Data Processing
dask[complete]==2024.1.0
pyarrow==14.0.1



# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
folium==0.15.1

# Dashboard
streamlit==1.29.0

# Weather API
requests==2.31.0

# PDF Generation
reportlab==4.0.7
matplotlib==3.8.2

# Geospatial
geopandas==0.14.1
shapely==2.0.2